+++ core/app_system.go
@@ -87,7 +87,7 @@ func (app *application) sysGetApplicationOrganization(ID string) (*model.Applica
 }
 
 func (app *application) sysCreateApplicationOrganization(appOrg model.ApplicationOrganization, appID string, orgID string) (*model.ApplicationOrganization, error) {
-	application, err := app.storage.FindApplication(appID)
+	application, err := app.storage.FindApplication(nil, appID)
 	if err != nil || application == nil {
 		return nil, errors.WrapErrorData(logutils.StatusInvalid, model.TypeApplication, nil, err)
 	}
@@ -169,7 +169,7 @@ func (app *application) sysUpdateOrganization(ID string, name string, requestTyp
 }
 
 func (app *application) sysGetApplication(ID string) (*model.Application, error) {
-	appAdm, err := app.storage.FindApplication(ID)
+	appAdm, err := app.storage.FindApplication(nil, ID)
 	if err != nil {
 		return nil, errors.WrapErrorAction(logutils.ActionGet, model.TypeApplication, nil, err)
 	}
@@ -184,16 +184,78 @@ func (app *application) sysCreateApplication(name string, multiTenant bool, admi
 	now := time.Now()
 
 	// application
-
-	applicationID, _ := uuid.NewUUID()
-	application := model.Application{ID: applicationID.String(), Name: name, MultiTenant: multiTenant, Admin: admin, SharedIdentities: sharedIdentities,
+	for i, at := range appTypes {
+		appTypes[i].ID = uuid.NewString()
+		appTypes[i].DateCreated = now
+		for vidx := range at.Versions {
+			appTypes[i].Versions[vidx].ID = uuid.NewString()
+			appTypes[i].Versions[vidx].ApplicationType = model.ApplicationType{ID: appTypes[i].ID}
+			appTypes[i].Versions[vidx].DateCreated = now
+		}
+	}
+	application := model.Application{ID: uuid.NewString(), Name: name, MultiTenant: multiTenant, Admin: admin, SharedIdentities: sharedIdentities,
 		Types: appTypes, DateCreated: now}
 
 	insertedApplication, err := app.storage.InsertApplication(nil, application)
 	if err != nil {
-		return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeApplication, nil, err)
+		return nil, errors.WrapErrorAction(logutils.ActionCreate, model.TypeApplication, nil, err)
 	}
 	return insertedApplication, nil
+}
+
+func (app *application) sysUpdateApplication(ID string, name string, multiTenant bool, admin bool, sharedIdentities bool, appTypes []model.ApplicationType) error {
+	transaction := func(context storage.TransactionContext) error {
+		//1. find application
+		application, err := app.storage.FindApplication(context, ID)
+		if err != nil {
+			return errors.WrapErrorAction(logutils.ActionFind, model.TypeApplication, nil, err)
+		}
+		if application == nil {
+			return errors.ErrorData(logutils.StatusMissing, model.TypeApplication, nil)
+		}
+
+		//2. update app type list
+		updated := false
+		now := time.Now().UTC()
+		for i, at := range appTypes {
+			existingAppType := application.FindApplicationType(at.Identifier)
+			if existingAppType != nil {
+				//unchanged app type identifier, so set existing ID
+				appTypes[i].ID = existingAppType.ID
+				appTypes[i].Versions = existingAppType.Versions
+				appTypes[i].DateCreated = existingAppType.DateCreated
+				if at.Name != existingAppType.Name {
+					appTypes[i].DateUpdated = &now
+					updated = true
+				}
+			} else {
+				//added app type identifier, so set new ID
+				appTypes[i].ID = uuid.NewString()
+				for vidx := range at.Versions {
+					appTypes[i].Versions[vidx].ID = uuid.NewString()
+					appTypes[i].Versions[vidx].ApplicationType = model.ApplicationType{ID: appTypes[i].ID}
+					appTypes[i].Versions[vidx].DateCreated = now
+				}
+				appTypes[i].DateCreated = now
+				updated = true
+			}
+		}
+
+		//3. update if app types or other application params were updated
+		updated = updated || (name != application.Name) || (multiTenant != application.MultiTenant) || (admin != application.Admin) || (sharedIdentities != application.SharedIdentities)
+		if updated {
+			updatedApp := model.Application{ID: application.ID, Name: name, MultiTenant: multiTenant, Admin: admin, SharedIdentities: sharedIdentities,
+				Types: appTypes, DateCreated: application.DateCreated, DateUpdated: &now}
+			err = app.storage.SaveApplication(context, updatedApp)
+			if err != nil {
+				return errors.WrapErrorAction(logutils.ActionSave, model.TypeApplication, nil, err)
+			}
+		}
+
+		return nil
+	}
+
+	return app.storage.PerformTransaction(transaction)
 }
 
 func (app *application) sysGetApplications() ([]model.Application, error) {