+++ core/mocks/Storage.go
@@ -537,13 +537,13 @@ func (_m *Storage) FindAppOrgRolesByIDs(context storage.TransactionContext, ids 
 	return r0, r1
 }
 
-// FindApplication provides a mock function with given fields: ID
-func (_m *Storage) FindApplication(ID string) (*model.Application, error) {
-	ret := _m.Called(ID)
+// FindApplication provides a mock function with given fields: context, ID
+func (_m *Storage) FindApplication(context storage.TransactionContext, ID string) (*model.Application, error) {
+	ret := _m.Called(context, ID)
 
 	var r0 *model.Application
-	if rf, ok := ret.Get(0).(func(string) *model.Application); ok {
-		r0 = rf(ID)
+	if rf, ok := ret.Get(0).(func(storage.TransactionContext, string) *model.Application); ok {
+		r0 = rf(context, ID)
 	} else {
 		if ret.Get(0) != nil {
 			r0 = ret.Get(0).(*model.Application)
@@ -551,8 +551,8 @@ func (_m *Storage) FindApplication(ID string) (*model.Application, error) {
 	}
 
 	var r1 error
-	if rf, ok := ret.Get(1).(func(string) error); ok {
-		r1 = rf(ID)
+	if rf, ok := ret.Get(1).(func(storage.TransactionContext, string) error); ok {
+		r1 = rf(context, ID)
 	} else {
 		r1 = ret.Error(1)
 	}
@@ -1181,6 +1181,20 @@ func (_m *Storage) RemoveAccountsGroup(context storage.TransactionContext, group
 	var r0 error
 	if rf, ok := ret.Get(0).(func(storage.TransactionContext, string, []string) error); ok {
 		r0 = rf(context, groupID, accountIDs)
+	} else {
+		r0 = ret.Error(0)
+	}
+
+	return r0
+}
+
+// SaveApplication provides a mock function with given fields: context, application
+func (_m *Storage) SaveApplication(context storage.TransactionContext, application model.Application) error {
+	ret := _m.Called(context, application)
+
+	var r0 error
+	if rf, ok := ret.Get(0).(func(storage.TransactionContext, model.Application) error); ok {
+		r0 = rf(context, application)
 	} else {
 		r0 = ret.Error(0)
 	}